+++ src/Collada.cpp
@@ -1,3 +1,4 @@
+#include <math.h>
 
 #include "Collada.h"
