+++ tools/XmlConverter/main.cpp
@@ -72,7 +72,7 @@ Press Enter to continue.\n");
 					if (toLowerCase(nameb).find("model") != std::string::npos)				//it's a model
 					{
 						Swr_Model* model = new Swr_Model();
-						model->save_Xml(filename +"\\"+ filenameb, false);
+						model->save_Xml(filename +"/"+ filenameb, false);
 						delete model;
 					}
 				}
